requirements.txt
fastapi
uvicorn
pandas
matplotlib
reportlab
pywhatkit
yagmail
python-dotenv
supabase
PyJWT
bcrypt
python-multipart
requests